% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteFunction.R
\name{bloom_palette}
\alias{bloom_palette}
\title{palette generator}
\usage{
bloom_palette(name, type, n)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Episode1}, \code{Episode2},  \code{Episode3},
\code{Episode4}, \code{Episode5},  \code{Episode6},
\code{Episode7}, \code{Finale},
@importFrom graphics rgb rect par image text}

\item{type}{Type of palette desired. choices are:
\code{seq}, \code{qual},  \code{div}}

\item{n}{Number of colors desired.
Qual palettes have n=6, Seq palettes have n=8
All schemes are derived from Yan & Henck's outfits in The Big Flower Flight
If omitted, uses all colours.}
}
\value{
A vector of colours.
}
\description{
These are some colour palettes from a collection
}
\examples{
bloom_palette("Episode1", type="qual")
bloom_palette("Episode2", type="seq")
}
\keyword{colors}
